restructure: move the PostDetail to the "features" file
