+++ src/featuers/posts/PostDetail.js
@@ -2,6 +2,7 @@ import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import moment from "moment";
 import { grey } from "@mui/material/colors";
+import { KeyboardArrowLeft, KeyboardArrowRight } from "@mui/icons-material";
 import {
   Box,
   Card,
@@ -12,17 +13,96 @@ import {
   Stack,
   CircularProgress,
   Paper,
+  IconButton,
 } from "@mui/material";
-import MyCarousel from "./Carousel";
+
+import "react-responsive-carousel/lib/styles/carousel.min.css";
+import { Carousel } from "react-responsive-carousel";
+
 import { useDispatch, useSelector } from "react-redux";
 import { getPost } from "../../actions/postsAction";
-//1. background image
-//2. image gallary
-//3. acount details
+
+const ImageCarousel = ({ children }) => {
+  const arrowStyles = {
+    position: "absolute",
+    zIndex: 3,
+    top: "calc(50% - 40px)",
+    width: 40,
+    height: 40,
+    cursor: "pointer",
+    backgroundColor: "#2196f3",
+    color: "#fff",
+  };
+
+  const indicatorStyles = {
+    background: "#fff",
+    width: 12,
+    height: 12,
+    display: "inline-block",
+    margin: "0 12px",
+    borderRadius: "50%",
+    cursor: "pointer",
+  };
+  return (
+    <Carousel
+      dynamicHeight={true}
+      showThumbs={false}
+      renderArrowPrev={(onClickHandler, hasPrev, label) =>
+        hasPrev && (
+          <IconButton
+            type="button"
+            onClick={onClickHandler}
+            title={label}
+            style={{ ...arrowStyles, left: 20 }}
+          >
+            <KeyboardArrowLeft />
+          </IconButton>
+        )
+      }
+      renderArrowNext={(onClickHandler, hasNext, label) =>
+        hasNext && (
+          <IconButton
+            type="button"
+            onClick={onClickHandler}
+            title={label}
+            style={{ ...arrowStyles, right: 20 }}
+          >
+            <KeyboardArrowRight />
+          </IconButton>
+        )
+      }
+      renderIndicator={(onClickHandler, isSelected, index, label) => {
+        if (isSelected) {
+          return (
+            <span
+              style={{ ...indicatorStyles, background: "#2196f3" }}
+              aria-label={`Selected: ${label} ${index + 1}`}
+              title={`Selected: ${label} ${index + 1}`}
+            />
+          );
+        }
+        return (
+          <span
+            style={{ ...indicatorStyles }}
+            onClick={onClickHandler}
+            onKeyDown={onClickHandler}
+            value={index}
+            key={index}
+            role="button"
+            tabIndex={0}
+            title={`${label} ${index + 1}`}
+            aria-label={`${label} ${index + 1}`}
+          />
+        );
+      }}
+    >
+      {children}
+    </Carousel>
+  );
+};
 
 const PostDetails = () => {
   const { id } = useParams();
-
   const dispatch = useDispatch();
   const { post, isLoading } = useSelector((state) => state.posts);
 
@@ -57,7 +137,7 @@ const PostDetails = () => {
               display: "block",
             }}
           >
-            <MyCarousel>
+            <ImageCarousel>
               {post &&
                 post.image.map((element, index) => (
                   <div key={index} style={{ backgroundColor: "#304352" }}>
@@ -68,7 +148,7 @@ const PostDetails = () => {
                     />
                   </div>
                 ))}
-            </MyCarousel>
+            </ImageCarousel>
           </Box>
 
           {/* post message section + users details  section */}
